Refactor utility functions and tests
