--- src/graph/utils.ts
@@ -1,8 +1,0 @@
-import { v4 as uuid } from 'uuid';
-
-function* generateTokenId(): IterableIterator<string> {
-    yield uuid();
-}
-
-
-
